Remove the obsolete WizLab planning section
This was a list of incomplete ideas used while the portal was
developed. I'd thought about moving it to the dev wiki, but none of the
ideas here are very detailed, and some refer to spells or items that
have been removed.

[skip ci]

(cherry picked from commit 092858d3ad986602ce60a04a438dfbc31e5eb559)

Conflicts:
	crawl-ref/source/dat/des/portals/wizlab.des